more stupid macros gone. moved util functions from plibs to putil
