+++ pclsync/prun.c
@@ -34,7 +34,7 @@ static void start_thread(const char *name, void *run, void *ptr) {
   pthread_t thread;
   pthread_attr_t attr;
   
-  data = psync_new(thread_data);
+  data = malloc(sizeof(thread_data));
   data->run = run;
   data->ptr = ptr;
   data->name = name;
