+++ pclsync/ptimer.c
@@ -193,7 +193,7 @@ psync_timer_t ptimer_register(psync_timer_callback func, time_t numsec,
   psync_timer_t timer;
   uint32_t i;
   time_t n;
-  timer = psync_new(psync_timer_structure_t);
+  timer = malloc(sizeof(psync_timer_structure_t));
   timer->call = func;
   timer->param = param;
   n = TIMER_ARRAY_SIZE;
@@ -242,7 +242,7 @@ int ptimer_stop(psync_timer_t timer) {
 
 void ptimer_exception_handler(psync_exception_callback func) {
   struct exception_list *t;
-  t = psync_new(struct exception_list);
+  t = malloc(sizeof(struct exception_list));
   t->next = NULL;
   t->func = func;
   t->threadid = pthread_self();
@@ -254,7 +254,7 @@ void ptimer_exception_handler(psync_exception_callback func) {
 
 void ptimer_sleep_handler(psync_exception_callback func) {
   struct exception_list *t;
-  t = psync_new(struct exception_list);
+  t = malloc(sizeof(struct exception_list));
   t->next = NULL;
   t->func = func;
   t->threadid = pthread_self();
